main: improve action logging
this is both to improve user friendlyness and make the output of the
rust binary more like the output of the old shell script it is
replacing. now we print out more information about what is happening,
including the fingerprints of the keys it is affecting. we also now note
that the authorized_keys file is automatically generated in a comment.