tests: modify test script to test options
lets make sure that these are always supported by having a particularly
difficult case in the tests. as long as one passes, we are probably
fine, since most of the heavy lifting is done by the openssh-keys
library and that already has tests for all the edge cases. we just need
to make sure it works at all.