lib: warn and omit invalid public keys
when update-ssh-keys finds an invalid key in the authorized_keys
directory, it would throw an error and quit. this is a different
behavior than the shell script, which didn't do any validation on the
contents of the authorized_keys directory. this commit reimplements the
read_keys function that is in the openssh-keys library with a slight
change; instead of returning an error when the parse fails, it prints
out a warning and omits the key from authorized keys. this is also a
slightly different behavior, but it brings in line the validation
present in this version with the not-crashing of the shell script.

this implementation adds an AuthorizedKeyEntry sum type which stores
both Valid and Invalid keys as either parsed PublicKey or String types
respectively. this allows different parts of the program to behave
differently with regard to the validity of the key. in particular, the
write function, which updates the authorized_keys directory, writes the
keys regardless of validity, so as to not delete keys out underneath
users. the sync function, which updates the authorized_keys file, only
emits valid keys.